+++ LibraryComponents/Classes/Core/AndesDropdown/Config/AndesDropdownViewConfigFactory.swift
@@ -10,17 +10,18 @@ import Foundation
 internal class AndesDropdownViewConfigFactory {
     static func provide(from dropdown: AndesDropdown, isSelected: Bool) -> AndesDropdownViewConfig {
 
-        if let style = AndesDropdownStyleFactory.getStyle(dropdown.triggerType) {
+        let triggerType = dropdown.triggerType
+        let style = AndesDropdownStyleFactory.getStyle(triggerType.type)
+        switch triggerType.type {
+        case .formDropdown:
             return AndesDropdownViewConfig(borderColor: style.borderColor,
-                                           counter: dropdown.counter,
-                                           placeholderText: dropdown.placeholder,
+                                           placeholderText: (dropdown.triggerType as? FormDropdownTrigger)?.placeholder,
                                            icon: isSelected ? AndesIcons.chevronUp20 : AndesIcons.chevronDown20,
-                                           title: dropdown.title)
+                                           title: triggerType.title)
+        case .standalone:
+            return AndesDropdownViewConfig(borderColor: style.borderColor,
+                                           icon: isSelected ? AndesIcons.chevronUp20 : AndesIcons.chevronDown20,
+                                           title: triggerType.title)
         }
-        return AndesDropdownViewConfig(borderColor: UIColor.Andes.gray250,
-                                       counter: dropdown.counter,
-                                       placeholderText: dropdown.placeholder,
-                                       icon: isSelected ? AndesIcons.chevronUp20 : AndesIcons.chevronDown20,
-                                       title: dropdown.title)
     }
 }