+++ LibraryComponents/Classes/Core/AndesDropdown/Controller/Config/AndesDropdownBottomSheetCellFactory.swift
@@ -9,15 +9,15 @@ import Foundation
 
 class AndesDropdownBottomSheetCellFactory {
 
-    static func provide(menuCellType: [AndesDropDownMenuCellType], size: AndesListSize, numberOfLines: Int) -> [AndesSimpleCell] {
+    static func provide(menuCellType: [AndesDropDownMenuCell], size: AndesListSize, numberOfLines: Int) -> [AndesSimpleCell] {
         var cell: [AndesSimpleCell] = []
         menuCellType.forEach({ (cellType) in
             cell.append(AndesSimpleCell(withTitle: cellType.title,
-                                        size: size,
                                         subtitle: nil,
                                         thumbnail: cellType.thumbnail,
                                         numberOfLines: numberOfLines))
         })
+
         return cell
     }
 }