+++ LibraryComponents/Classes/Core/AndesDropdown/Type/Menu/AndesDropdownMenuType.swift
@@ -2,17 +2,19 @@
 //  AndesDropdownMenuType.swift
 //  AndesUI
 //
-//  Created by Jonathan Alonso Pinto on 13/11/20.
+//  Created by Jonathan Alonso Pinto on 27/11/20.
 //
 
 import Foundation
 
 @objc public enum AndesDropdownMenuType: Int, AndesEnumStringConvertible {
-    case bottomSheet
+    case botttomSheet
+    case floatingMenu
 
     public static func keyFor(_ value: AndesDropdownMenuType) -> String {
         switch value {
-        case .bottomSheet: return " BOTTOMSHEET"
+        case .botttomSheet: return "BOTTOMSHEET"
+        case .floatingMenu: return "FLOATINGMENU"
         }
     }
 }