+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### 🛠 Bug fixes
 - AndesTagChoice text color changed. | Authors: [@JBAZANCARRIZ](https://github.com/JBAZANCARRIZ)
 - AndesList fix border in thumbnail | Authors: [@joalonsopint](https://github.com/joalonsopint)
+- AndesList fix removing cell size | Authors: [@joalonsopint](https://github.com/joalonsopint)
 
 ### 🚀 Features
 - Date Picker component | Authors: [@AndriuCoelho](https://github.com/AndriuCoelho)