+++ CHANGELOG.md
@@ -1,8 +1,11 @@
+# v3.15.0
+### 🛠 Bug fixes
+- AndesList fix removing cell size | Authors: [@joalonsopint](https://github.com/joalonsopint)
+- AndesList fix border in thumbnail | Authors: [@joalonsopint](https://github.com/joalonsopint)
+
 # v3.14.1
 ### 🛠 Bug fixes
 - AndesTagChoice text color changed. | Authors: [@JBAZANCARRIZ](https://github.com/JBAZANCARRIZ)
-- AndesList fix border in thumbnail | Authors: [@joalonsopint](https://github.com/joalonsopint)
-- AndesList fix removing cell size | Authors: [@joalonsopint](https://github.com/joalonsopint)
 
 ### 🚀 Features
 - Date Picker component | Authors: [@AndriuCoelho](https://github.com/AndriuCoelho)